Update dot11decrypt.c
fix err:
/Users/runner/work/wireshark/wireshark/epan/crypt/dot11decrypt.c:1657:45: error: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'INT' (aka 'int') [-Werror,-Wshorten-64-to-32]
                    key_index = ctx->keys_nr++;
                              ~ ~~~~~~~~~~~~^~
1 error generated.

+++ epan/crypt/dot11decrypt.c
@@ -1652,9 +1652,9 @@ Dot11DecryptRsna4WHandshake(
                 }
 
                 if(ctx->keys_nr < 255) {
-                    key_index = ctx->keys_nr++;
+                    key_index = (INT)ctx->keys_nr++;
                 } else {
-                    static int s_key_index = 0;
+                    static INT s_key_index = 0;
                     key_index = s_key_index++;
                 }
